+++ ouroboros-network-api/src/Ouroboros/Network/PeerSelection/LedgerPeers/Type.hs
@@ -4,6 +4,7 @@
 module Ouroboros.Network.PeerSelection.LedgerPeers.Type
   ( PoolStake (..)
   , AccPoolStake (..)
+  , IsBigLedgerPeer (..)
   ) where
 
 import           Control.DeepSeq (NFData (..))
@@ -21,3 +22,18 @@ newtype PoolStake = PoolStake { unPoolStake :: Rational }
 --
 newtype AccPoolStake = AccPoolStake { unAccPoolStake :: Rational }
     deriving (Eq, Fractional, Num, Ord, Show)
+
+
+-- | A boolean like type.  Big ledger peers are the largest SPOs which control
+-- 90% of staked stake.
+--
+-- Note that 'IsBigLedgerPeer' indicates a role that peer plays in the eclipse
+-- evasion, e.g. that a peer was explicitly selected as a big ledger peer, e.g.
+-- 'IsNotBigLedgerPeer' does not necessarily mean that the peer isn't a big
+-- ledger peer.  This is because we select root peers from all ledger peers
+-- (including big ones).
+--
+data IsBigLedgerPeer
+   = IsBigLedgerPeer
+   | IsNotBigLedgerPeer
+  deriving Eq