+++ ouroboros-network-api/CHANGELOG.md
@@ -1,5 +1,14 @@
 # Revision history for ouroboros-network-api
 
+## next version
+
+### Breaking changes
+
+### Non-breaking changes
+
+* Added `IsBigLedgerPeer` type to
+  `Ouroboros.Network.PeerSelection.LedgerPeers.Type`.
+
 ## 0.5.0.0 -- 2023-05-15
 
 * Swapped `NodeToClientV_15` with `NodeToClientV_16`, e.g. handshake query
@@ -37,7 +46,7 @@ Depracated release, use `0.5` instead.
   - New PeerSharing data type file
 - Remove foldr in favour of toOldestFirst
 
-### Non breaking
+### Non-breaking
 
 ## 0.1.0.0 -- 2022-11-17